Changes to move the storybook build to the workspace root
- moving the SB build script to root
- Extending the main tsconfig to reuse for SB pkg
- deleting unused files
- fixing tsc compile errors for @glimmerx/storybook pkg
