+++ packages/@glimmerx/storybook/.npmignore
@@ -1,0 +1,1 @@
+.babelrc*
