+++ packages/@glimmerx/storybook/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../../tsconfig",
+  "compilerOptions": {
+    "noEmit": true
+  },
+  "exclude": [],
+  "include": [
+    "src/**/*.ts"
+  ]
+
+}
